Componentization of the 'signature' module.
Similar to 279ca8e1b0e0f4eef93c0c5ab458671b095ec41f.

I agree to license my contributions to each file under the
same terms given at the top of each file I changed.
